chore: removed unused constructor and Initialise function arguments from ANSYS models and fixed all related calls (in as much as i have access to) chore: refactored Model(topology,vset) constructor ascertaining that only the highest dim elements of same-named regions are kept refactored: boundary_face_vec_ creation method so that it excludes line elements that may be contained inside a modelsubdomain because these cannot have boundary faces chore: fixed Region_Test that failed on a boundary face consistency check

+++ source/includes/model/ModelSubDomain.h
@@ -97,7 +97,7 @@ class ModelSubDomain {
     void SortVectors( size_t interior_cells, size_t interior_nodes );
     
     /// assuming that a partitioned (and sorted) cell vector is in place, constructs the bd_face_vec_ by checking whether neighbor cells belong to the domain or not
-    void BuildPerimeterFaceVector( size_t interior_cells );
+    void BuildPerimeterFaceVector( int64_t interior_cells );
     
     /// removes any cells or node pointers that were set to zero elsewhere; returns number of cells removed
     size_t RemoveNullPointerCells();
